backend/requirements.txt
fastapi==0.104.1
pypdf==3.16.0
uvicorn==0.23.2
python-dotenv==1.0.0
pydantic==2.4.2
faiss-cpu==1.7.4
numpy==1.24.3
pandas==2.0.3
python-multipart==0.0.6
docx2txt
torch>=2.0.0
langchain==0.0.335
langchain-community==0.0.13
langchain-openai
langchain_experimental
huggingface-hub==0.26.0
psycopg2-binary==2.9.9
sqlalchemy==2.0.23
uuid==1.30
pymilvus==2.4.9
milvus-model==0.2.8
redis
FlagEmbedding==1.2.11